backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Database & Storage
chromadb>=0.4.18
sqlite3

# Web Scraping
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# AI/ML
sentence-transformers>=2.2.0
torch>=2.0.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0

# LLM Integration (Optional)
openai>=1.0.0
